requirements.txt
git+https://github.com/tpiscitell/kurator.git#egg=kurator-master
kafka-python==0.9.2
kazoo==2.0
scapy>=2.2.0
